+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0 (2024-07-23)
 
 - Expects `toHaveBeenNthCalledWith` followed by `toHaveBeenCalledTimes`
+- Different error message for `toHaveBeenNthCalledWith`
 
 ## 1.0.0 (2024-07-23)